requirements.txt
crewai==0.61.0
crewai-tools==0.8.3
python-dotenv==1.0.1
openai==1.54.0
requests==2.32.3
streamlit==1.38.0
fpdf2==2.8.1
python-docx==0.8.11
langchain-openai==0.0.1
pysqlite3-binary==0.5.0
duckdb>=0.7.1
# pysqlite3==0.4.7